build: Use go modules
We currently use glide for package management and it has several
problems associated with it such as with glide we have to make sure
we are in a go path which is sometime very frustrating to do. As recent
versions fo go now supports go modules we should go with them as they
allow the flexibility to place our code even outside a GOPATH. Also
other projects like nebraska already use go mods. Thus this patch inits
the project with go mods.