Add simple benchmark subcommand
The "instance fake" subcommand creates stress for the server but
doesn't give results to compare two servers in terms of how many
responses they sent out.
Add a simple "instance bench" subcommand that takes the same parameters
plus an optional benchmark duration which defaults to "--seconds 10".
The benchmark load does not consist of forced update circles of a
couple of instances but rather of single update requests from many new
unique instances. This is done to prevent instances entering a strange
state or encountering a rate limit because the server detects it as
hammering and refuses to process the request. The benchmark result is
very limited as it only shows the average number of responses got per
second. Saving a distribution of response times would give more
reliable results.